feat: add concurrency to workflow
Allow only one workflow at that time.

If a second workflow is triggered, it will be put in pending state, and
only start once the first workflow finish.
If the workflow is triggered a third time while the first one is
running, the second workflow is canceled, and the third takes over the
pending state.

Fix #163.

Signed-off-by: Alejandro Lazaro <alazaro@werpo.com.ar>