Allow more flexibility in detected provider release assetst
This allows the filenames of release assets to include the "v" version
prefix, which is decidedly non-standard.

When run against the whole registry, it only produces changes in providers/m/microsoft/azuredevops.json
and providers/u/usenimbus/nimbus.json

This code path should not be commonly used and we should reach out to
those provider authors to request they fix their build process instead
of relying on this workaround.

Signed-off-by: Christian Mesh <christianmesh1@gmail.com>